don't delete cached files when the server says you can use them :-o
darcs-hash:20050226055901-2ee97-052e85cf626de6d6f3b4e2694a5ec525d78ed2fa

+++ org/fenfire/fenfeed/http/HTTPResource.java
@@ -138,10 +138,12 @@ public final class HTTPResource {
 	boolean changed;
 	newURI = null;
 
-	file("content").delete();
-	file("location").delete();
-	file("header").delete();
-	file("lastRead").delete();
+	if(code != 304) {
+	    file("content").delete();
+	    file("location").delete();
+	    file("header").delete();
+	    file("lastRead").delete();
+	}
 
 
 	if(code == 200) {
